requirements.txt
alembic==1.14.1
aniso8601==9.0.1
atomicwrites==1.4.1
attrs==24.3.0
blinker==1.9.0
click==8.1.3
colorama==0.4.6
Flask==2.1.3
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==3.1.1
gunicorn==20.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.8
MarkupSafe==2.1.1
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
py==1.11.0
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==7.1.2
python-dotenv==1.0.1
pytz==2022.1
setuptools==75.8.0
six==1.16.0
SQLAlchemy==1.4.45
tomli==2.2.1
typing_extensions==4.12.2
Werkzeug==2.2.0
